+++ src/main/js/wiki/containers/Wiki.tsx
@@ -6,6 +6,7 @@ import Search from "./Search";
 import Page from "./Page";
 import Navigation from "../../Navigation";
 import WikiRoot from "./WikiRoot";
+import Settings from "./Settings";
 
 type Params = {
   repository: string;
@@ -28,6 +29,7 @@ const Wiki: FC<Props> = (props) => {
       <Switch>
         <Route exact path="/:repository/:branch" component={WikiRoot} />
         <Route path="/:repository/:branch/pages" component={Directory} />
+        <Route path="/:repository/:branch/settings" component={Settings} />
         <Route path="/:repository/:branch/history" component={History} />
         <Route path="/:repository/:branch/search" component={Search} />
         <Route path="/:repository/:branch" component={Page} />