+++ src/main/js/wiki/hooks/wiki.ts
@@ -1,4 +1,4 @@
-import { useQuery, UseQueryResult } from "react-query";
+import { useMutation, UseMutationResult, useQuery, useQueryClient, UseQueryResult } from "react-query";
 import { apiClient } from "../../apiclient";
 import { WikiDto } from "../types/wikiDto";
 
@@ -7,5 +7,46 @@ export function useWiki(repository: string, branch: string, refetchOnFocus = tru
     ["wiki", { repository: repository, branch: branch }],
     () => apiClient.get(`/repositories/${repository}/branches/${branch}`).then((response) => response.json()),
     { refetchOnWindowFocus: refetchOnFocus }
+  );
+}
+
+type editParams = {
+  rootDir: string;
+  landingPage: string;
+};
+
+export function useEditWiki(repository: string, branch: string): UseMutationResult {
+  const queryClient = useQueryClient();
+  const url = `/repositories/${repository}/branches/${branch}`;
+  return useMutation(
+    ["editWiki", url],
+    (data: editParams) => {
+      return apiClient.patch(url, data);
+    },
+    {
+      onSuccess: () => {
+        queryClient.removeQueries(["wiki", { repository: repository, branch: branch }]);
+      }
+    }
+  );
+}
+
+export function useInitWiki(repository: string, branch: string): UseMutationResult {
+  const queryClient = useQueryClient();
+  const url = `/repositories/${repository}/branches/${branch}`;
+  const data = {
+    landingPage: "Home",
+    rootDir: "docs"
+  };
+  return useMutation(
+    ["initWiki", url],
+    () => {
+      return apiClient.post(url, data);
+    },
+    {
+      onSuccess: () => {
+        queryClient.removeQueries(["wiki", { repository: repository, branch: branch }]);
+      }
+    }
   );
 }