+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Option to init smeagol wiki #3
+* Settings menu to change wiki paths
 
 ## [v1.2.0-1] - 2021-04-14
 ### Added